My Working Build

+++ engine/lib_maths/maths.h
@@ -24,8 +24,8 @@ template <typename T> Vector2<T> normalize(const Vector2<T>& v) {
   double l = length(v);
 
   if (l != 0) {
-    vector.x = v.x / l;
-    vector.y = v.y / l;
+    vector.x = v.x / (float)l;
+    vector.y = v.y / (float)l;
   }
 
   return vector;
@@ -38,6 +38,14 @@ Vector2<T> operator*(const Vector2<T>& left, const Vector2<T>& right) {
   r.y *= right.y;
   return r;
 }
+
+//template <typename T>
+//Vector2<T> operator*(const Vector2<T>& left, const float& val) {
+//	Vector2<T> r = left;
+//	r.x *= val;
+//	r.y *= val;
+//	return r;
+//}
 
 template <typename T>
 std::ostream& operator<<(std::ostream& os, const Vector2<T>& v) {
@@ -65,7 +73,7 @@ Vector2<T> rotate(const Vector2<T>& v, const double degrees) {
 
 static int RandomInt(int min, int max)	// Generate random int between min[inclusive] and max[exclusive]
 {
-	return (((float)rand()) / ((float)RAND_MAX) * (max - min)) + min;
+	return (((float)rand() / (float)RAND_MAX) * (max - min)) + min;
 }
 
 static float RandomFloat(float min, float max)	// Generate random float between min[inclusive] and max[exclusive]
@@ -73,7 +81,7 @@ static float RandomFloat(float min, float max)	// Generate random float between 
 	return (((float)rand() / (float)RAND_MAX) * (max - min)) + min;
 }
 
-static float RandomDouble(double min, double max)	// Generate random double between min[inclusive] and max[exclusive]
+static double RandomDouble(double min, double max)	// Generate random double between min[inclusive] and max[exclusive]
 {
 	return (((double)rand() / (double)RAND_MAX) * (max - min)) + min;
 }
